Basic functionality for creating actions with sensors
