+++ HomeApp/app/src/main/java/com/HomeApp/screens/SensorActionScreen.kt
@@ -1,0 +1,60 @@
+package com.HomeApp.screens
+
+import android.util.Log
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.FabPosition
+import androidx.compose.material.Scaffold
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.ArrowForward
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.HomeApp.ui.composables.ActionCard
+import com.HomeApp.ui.composables.RoutinesFAB
+import com.HomeApp.ui.composables.RoutinesTitleBar
+import com.HomeApp.ui.composables.RoutinesTitleBarItem
+import com.HomeApp.ui.navigation.SensorFinish
+
+@Composable
+fun SensorActionScreen(
+    navController: NavController,
+    OnSelfClick: () -> Unit = {}
+) {
+    val isDevices = SelectedItems.getType()[0]
+    val listHeight = LocalConfiguration.current.screenHeightDp
+    val document = SelectedItems.getItems()[0]
+
+    Scaffold(
+        topBar = {
+            RoutinesTitleBar(
+                item = RoutinesTitleBarItem.SensorActions,
+                navController = navController
+            )
+        },
+        content = {
+            LazyColumn(
+                modifier = androidx.compose.ui.Modifier
+                    .height(listHeight.dp)
+                    .padding(it)
+                    .padding(vertical = 10.dp, horizontal = 20.dp),
+                verticalArrangement = Arrangement.spacedBy(10.dp),
+                content = {
+                    item { ActionCard(document = document) }
+                }
+            )
+        },
+        floatingActionButton = {
+            RoutinesFAB(
+                icon = Icons.Rounded.ArrowForward,
+                onClick = { navController.navigate(SensorFinish.route) }
+            )
+        },
+        isFloatingActionButtonDocked = true,
+        floatingActionButtonPosition = FabPosition.End
+    )
+}
